Computing corr

This function computes the weighted correlation with a penalty for lags. It should only be used after the fixed lags have already been applied to the dataset and timepoints using the functions prep.data() and best.lag().

comp.corr(data, time, C)

Arguments

data
a lagged matrix or data frame with rows representing genes and columns representing different timepoints (NAs added when lags are needed)

time
a lagged matrix with rows representing each gene's timepoint and columns representing the number of timepoints, NA is introduced when it is lagged

C
a numeric value of C used in computing weighted correlation

Value

a simmilarity matrix with values between -1 and 1 (1 highly correlated, 0 no correlation)

Examples

## This function computes the correlation after the lags (or shifts) have
## been computed.  In this example, the lags argument is randomly sampled
## for the sake of illustrating how prep.data() applies the lags and
## prepares a transformed dataset for comp.corr().
lagged <- prep.data(array(rnorm(30), c(3, 10)), timepoints = seq(0, 45, 5),
          lags = sample(c(0, 1, -1, 2, -2), size = 3))
comp.corr(data = lagged$data, time = lagged$time, C = 10)

#>            1          2
#> 2 -0.1524207           
#> 3  0.2619629  0.1669154

## This example shows how comp.corr is used in practice with real data.
## The best.lag() function is called first to pre-compute the lags, which
## are passed to prep.data().
randdata <- array(rnorm(120), c(10, 12))
bl <- best.lag(data = randdata, timepoints = 1:12, C = 5)
lag.data <- prep.data(randdata, timepoints = 1:12, lags = bl)
comp.corr(lag.data$data, time = lag.data$time, C = 5)

#>              1           2           3           4           5           6
#> 2  -0.35822127                                                            
#> 3  -0.15418123  0.10934719                                                
#> 4   0.87356260 -0.58680190 -0.11289677                                    
#> 5  -0.36068706 -0.11587157 -0.60325691 -0.26271861                        
#> 6  -0.28300733  0.26714178  0.27042798 -0.26757381 -0.13654873            
#> 7  -0.38303289 -0.31864939  0.29125844 -0.09353025  0.35404957  0.12830633
#> 8   0.04984673  0.17401303  0.53082994 -0.01647289 -0.58374117  0.19810114
#> 9   0.18477504 -0.36434806  0.25222517  0.23074225 -0.09000143 -0.13542134
#> 10  0.42484772  0.42371700  0.16453183  0.02985382 -0.52280737  0.01780643
#>              7           8           9
#> 2                                     
#> 3                                     
#> 4                                     
#> 5                                     
#> 6                                     
#> 7                                     
#> 8  -0.01488773                        
#> 9   0.35438023 -0.07792687            
#> 10 -0.65952337  0.40868636 -0.36892601