Computes best lag correlation

This function computes correlation based on best picked lags. The lags indicate delayed changes.

corr.bestlag(data, timepoints, max.lag = NULL, C = NULL, penalty = "high",
  iter = 10)

Arguments

data
a matrix with rows representing genes and columns representing different timepoints

timepoints
a vector of time points used in the dataset

max.lag
a integer value of the maximum lags allowed in the dataset, if null, defaults to the floor of the number of timepoints divided by 4

C
a numeric value of C used in computing weighted correlation, if null, a default is computed based on the penalty argument

penalty
a factor with two levels high and low penalty on the weighted correlation

iter
an integer indicating the number of C values to test for low penalty

Value

a list containing weighted correlation and best lags used in each row

Examples

corr.bestlag(array(rnorm(30), c(5, 6)), max.lag = 1,
          timepoints = c(0, 5, 10, 15, 20, 25), C = 10, penalty = "high")

#> $corr
#>            1          2          3          4
#> 2 -0.1517864                                 
#> 3  0.2248099 -0.5332766                      
#> 4  0.2976208 -0.8639861  0.1388098           
#> 5 -0.5652999  0.3731797 -0.6542605 -0.3130839
#> 
#> $lags
#> [1] 0 0 0 0 0
#> 
#> $C
#> [1] 10
#> 

corr.bestlag(array(runif(40, 0, 20), c(4, 10)),
          timepoints = c(0, 0.5, 1.5, 3, 6, 12, 18, 26, 39, 50), penalty = "high")

#> $corr
#>            1          2          3
#> 2 0.34642751                      
#> 3 0.16597141 0.04381154           
#> 4 0.56904909 0.34780431 0.33603452
#> 
#> $lags
#> [1]  0 -1  0  1
#> 
#> $C
#> [1] 145.7281
#> 

corr.bestlag(matrix(data = rexp(n = 40, 2), nrow = 8),
          timepoints = c(0, 5, 15, 20, 40), penalty = "low", iter = 5)

#> $corr
#>            1          2          3          4          5          6          7
#> 2 -0.7748456                                                                  
#> 3  0.5315331 -0.7485698                                                       
#> 4  0.2023999 -0.4808298  0.8632284                                            
#> 5 -0.7299764  0.2709059 -0.4995703 -0.3092350                                 
#> 6 -0.9027405  0.5488940 -0.4849267 -0.1617593  0.8589109                      
#> 7  0.4743248 -0.7656214  0.9869966  0.8930882 -0.3590924 -0.3540430           
#> 8 -0.9540451  0.9052804 -0.7861730 -0.4380442  0.6462569  0.8149260 -0.7068592
#> 
#> $lags
#> [1]  0  0  0 -1  1 -1  1  1
#> 
#> $C
#> [1] 2680.662
#>