Preparing Data

This function prepares the data to compute correlation by introducing NA's when lags are needed

prep.data(data, lags, timepoints)

Arguments

data
a matrix or data frame with rows representing genes and columns representing different timepoints. If data is a data frame, the gene names can be specified using the row.names().

lags
a vector of same length as the number of rows in the data column indicating the best lags

timepoints
a vector of time points used in the dataset

Value

a list of two matrices, one matrix with NA's for the lags for the dataset and another matrix with the timepoints used for each row in the dataset

Examples

prep.data(array(rnorm(20), c(5, 4)), c(0, 0, 0, -1, 1),
          timepoints = c(0, 5, 15, 30))

#> $data
#>             [,1]       [,2]       [,3]       [,4]
#> [1,] -0.05892959 -1.1262759  0.8137396  0.3243352
#> [2,] -0.53258072 -0.3694356 -1.4848223 -0.1388366
#> [3,]  1.07907227 -1.1815951 -1.4249221 -1.1836451
#> [4,]  0.05931762  1.9369219  1.6123666         NA
#> [5,]          NA -0.6660184 -1.8316199  0.6884226
#> 
#> $time
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    5   15   30
#> [2,]    0    5   15   30
#> [3,]    0    5   15   30
#> [4,]    5   15   30   NA
#> [5,]   NA    0    5   15
#> 

prep.data(array(runif(100, 0, 10), c(10, 10)), sample((-2:2), size = 10, replace = TRUE),
          timepoints = c(0, 5, 15, 30, 45, 60, 75, 80, 100, 120))

#> $data
#>           [,1]     [,2]     [,3]      [,4]     [,5]     [,6]      [,7]
#>  [1,]       NA 1.218646 6.229228 0.3857188 1.451154 6.305429 7.2628532
#>  [2,] 8.240329 9.671391 8.767677 1.6813245 9.821906 2.515776 0.3994066
#>  [3,] 6.230610 1.634123 2.859856 0.8603899 9.302815 2.264813 4.0527020
#>  [4,] 4.469974 5.138669 2.467509 2.4046123 4.194859 8.455401 2.6969463
#>  [5,] 7.082934 1.432664 3.979936 6.3888373 4.955599 3.458536 4.8828865
#>  [6,]       NA       NA 8.900290 6.5029870 6.320929 5.033951 3.5270469
#>  [7,]       NA       NA 1.043065 1.5529453 9.485788 2.505233 2.4873487
#>  [8,]       NA 6.304117 5.083099 3.4011924 9.790338 9.587279 1.1804457
#>  [9,] 6.273660 6.565796 3.599601 0.1369537 2.437913 5.697405 9.3642889
#> [10,] 8.249154 0.898489 1.989862 7.3572937 2.557493 6.964886 4.2943161
#>            [,8]      [,9]    [,10]
#>  [1,] 9.2396077 5.5231878 7.007675
#>  [2,] 7.7064318        NA       NA
#>  [3,] 6.2845017 4.3152240 6.164986
#>  [4,] 3.1611551 8.1423137 1.592747
#>  [5,] 9.2452039 5.6658628 4.015980
#>  [6,] 0.6731983 0.5106245 9.064917
#>  [7,] 0.2701945 0.0247683 0.919698
#>  [8,] 4.9583467 7.3779515 7.193180
#>  [9,] 7.9447767        NA       NA
#> [10,] 2.4258702        NA       NA
#> 
#> $time
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   NA    0    5   15   30   45   60   75   80   100
#>  [2,]   15   30   45   60   75   80  100  120   NA    NA
#>  [3,]    0    5   15   30   45   60   75   80  100   120
#>  [4,]    0    5   15   30   45   60   75   80  100   120
#>  [5,]    0    5   15   30   45   60   75   80  100   120
#>  [6,]   NA   NA    0    5   15   30   45   60   75    80
#>  [7,]   NA   NA    0    5   15   30   45   60   75    80
#>  [8,]   NA    0    5   15   30   45   60   75   80   100
#>  [9,]   15   30   45   60   75   80  100  120   NA    NA
#> [10,]   15   30   45   60   75   80  100  120   NA    NA
#>