Preparing Data

This function prepares the data to compute correlation by introducing NA's when lags are needed

prep.data(data, lags, timepoints)

Arguments

data
a matrix with rows representing genes and columns representing different timepoints

lags
a vector of same length as the number of rows in the data column indicating the best lags

timepoints
a vector of time points used in the dataset

Value

a list of two matrices, one matrix with NA's for the lags for the dataset and another matrix with the timepoints used for each row in the dataset

Examples

prep.data(array(rnorm(20), c(5, 4)), c(0, 0, 0, -1, 1),
          timepoints = c(0, 5, 15, 30))

#> $data
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.1100904  1.4172813  0.9721363 -1.4882559
#> [2,]  0.3720032 -0.8671011  0.2589025  2.4633691
#> [3,] -0.6098583 -2.2181411  1.2074691  0.5692981
#> [4,] -0.2753020 -0.2694864 -1.1107009         NA
#> [5,]         NA -0.3027063  0.3766451 -0.6983068
#> 
#> $time
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    5   15   30
#> [2,]    0    5   15   30
#> [3,]    0    5   15   30
#> [4,]    5   15   30   NA
#> [5,]   NA    0    5   15
#> 

prep.data(array(runif(100, 0, 10), c(10, 10)), sample((-2:2), size = 10, replace = TRUE),
          timepoints = c(0, 5, 15, 30, 45, 60, 75, 80, 100, 120))

#> $data
#>            [,1]      [,2]      [,3]      [,4]     [,5]      [,6]       [,7]
#>  [1,]        NA 2.5698292 8.5417365 7.2817997 4.968253 6.4850131 4.56749181
#>  [2,]        NA 6.1400560 0.7588305 9.7084502 0.147188 9.8456547 1.79034291
#>  [3,]        NA        NA 4.4149919 2.0582886 5.352176 9.5791912 5.36133257
#>  [4,]        NA        NA 3.1559469 4.9588965 1.664908 4.5308924 7.61059405
#>  [5,] 0.5817565 9.8153185 5.6438573 0.2488603 8.507213 0.4300129 5.59383239
#>  [6,]        NA        NA 2.7264620 4.1596145 4.591390 5.4470517 7.78454304
#>  [7,]        NA 6.5369940 4.3915000 6.2279882 3.600424 6.3906099 8.27921322
#>  [8,] 5.9481920 5.0314610 3.7896074 6.4306903 3.505882 9.4568785 9.82668667
#>  [9,] 0.2662467 0.7832827 7.4301824 0.7463363 3.595307 7.0848426 0.08658075
#> [10,] 0.1776996 5.2428013 9.6979189 8.4423142 8.707272 4.3089879 0.64883093
#>           [,8]     [,9]     [,10]
#>  [1,] 2.818126 7.249351 7.6211091
#>  [2,] 1.732248 4.122564 0.7838425
#>  [3,] 9.662092 1.096700 5.5086240
#>  [4,] 7.134120 5.862222 1.5636057
#>  [5,] 8.386300       NA        NA
#>  [6,] 6.264565 9.215262 6.8464411
#>  [7,] 6.756619 5.144025 3.4969548
#>  [8,] 5.786308       NA        NA
#>  [9,] 8.430969 4.088539 8.9790251
#> [10,] 3.608737 3.444290        NA
#> 
#> $time
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   NA    0    5   15   30   45   60   75   80   100
#>  [2,]   NA    0    5   15   30   45   60   75   80   100
#>  [3,]   NA   NA    0    5   15   30   45   60   75    80
#>  [4,]   NA   NA    0    5   15   30   45   60   75    80
#>  [5,]   15   30   45   60   75   80  100  120   NA    NA
#>  [6,]   NA   NA    0    5   15   30   45   60   75    80
#>  [7,]   NA    0    5   15   30   45   60   75   80   100
#>  [8,]   15   30   45   60   75   80  100  120   NA    NA
#>  [9,]    0    5   15   30   45   60   75   80  100   120
#> [10,]    5   15   30   45   60   75   80  100  120    NA
#>